+++ platformio.ini
@@ -16,9 +16,17 @@ build_flags = -Werror -Isrc/imxrt1062/MIMXRT1062 -Isrc/imxrt1062/CMSIS/Core/Incl
 platform = teensy
 board = teensy41
 framework = arduino
-build_src_filter = +<main.cpp> +<imxrt1062> +<teensy_sd.*>
+build_src_filter = +<imxrt1062> +<teensy_41>
 test_framework = unity
 test_filter = arduino/*
+
+[env:pal_9k31]
+platform = ststm32
+board = nucleo_g474re
+framework = stm32cube
+build_src_filter = +<stm32g4> +<pal_9k31>
+test_framework = unity
+test_filter = stm32cube/*
 
 [env:native]
 platform = native
